Adding method to get recent sensor records.

+++ app/src/main/java/com/nightscout/android/dexcom/records/SensorRecord.java
@@ -5,8 +5,8 @@ import java.nio.ByteOrder;
 
 public class SensorRecord extends GenericTimestampRecord {
 
-    private long unfiltered;
-    private long filtered;
+    private int unfiltered;
+    private int filtered;
     private int rssi;
     private int OFFSET_UNFILTERED = 8;
     private int OFFSET_FILTERED = 12;
